+++ frontend/src/components/listing/ListingMetadata.tsx
@@ -1,15 +1,7 @@
-import { useState } from "react";
-import {
-  FaArrowDown,
-  FaArrowUp,
-  FaClipboard,
-  FaEye,
-  FaUser,
-} from "react-icons/fa";
+import { FaClipboard, FaEye, FaUser } from "react-icons/fa";
 import { useNavigate } from "react-router-dom";
 
 import { useAlertQueue } from "@/hooks/useAlertQueue";
-import { useAuthentication } from "@/hooks/useAuth";
 import ROUTES from "@/lib/types/routes";
 
 interface Props {
@@ -20,66 +12,18 @@ interface Props {
   listingSlug: string | null;
   views: number;
   createdAt: number;
-  userVote: boolean | null;
 }
 
 const ListingMetadata = ({
-  listingId,
   creatorId,
   creatorName,
   creatorUsername,
   listingSlug,
   views,
   createdAt,
-  userVote: initialUserVote,
 }: Props) => {
-  const auth = useAuthentication();
-  const [voting, setVoting] = useState(false);
-  const [userVote, setUserVote] = useState(initialUserVote);
-  const { addErrorAlert, addAlert } = useAlertQueue();
+  const { addAlert } = useAlertQueue();
   const navigate = useNavigate();
-
-  const handleVote = async (vote: boolean) => {
-    setVoting(true);
-    try {
-      if (userVote === vote) {
-        const response = await auth.client.DELETE("/listings/{id}/vote", {
-          params: {
-            path: {
-              id: listingId,
-            },
-          },
-        });
-        if (response.error) {
-          addErrorAlert(response.error);
-        } else {
-          addAlert("Vote removed", "success");
-          setUserVote(null);
-        }
-      } else {
-        const response = await auth.client.POST("/listings/{id}/vote", {
-          params: {
-            path: {
-              id: listingId,
-            },
-            query: {
-              upvote: vote,
-            },
-          },
-        });
-        if (response.error) {
-          addErrorAlert(response.error);
-        } else {
-          addAlert("Vote added", "success");
-          setUserVote(vote);
-        }
-      }
-    } catch (error) {
-      addErrorAlert(error);
-    } finally {
-      setVoting(false);
-    }
-  };
 
   const listingTag = `${creatorUsername}/${listingSlug}`;
 
@@ -87,26 +31,6 @@ const ListingMetadata = ({
     <>
       {/* Metadata container - all items aligned left */}
       <div className="flex flex-wrap items-center gap-4 text-sm mb-2">
-        {/* Voting buttons */}
-        <div className="flex items-center gap-2 text-gray-600">
-          <button
-            className={`p-1 hover:bg-gray-100 rounded ${userVote === true && !voting ? "text-green-500" : ""}`}
-            onClick={() => handleVote(true)}
-            disabled={voting}
-          >
-            <FaArrowUp />
-          </button>
-          <button
-            className={`p-1 hover:bg-gray-100 rounded ${
-              userVote === false && !voting ? "text-red-500" : ""
-            }`}
-            onClick={() => handleVote(false)}
-            disabled={voting}
-          >
-            <FaArrowDown />
-          </button>
-        </div>
-
         {/* Creator button */}
         <div className="flex items-center gap-2">
           <button
